GSON Issues
So I guess trying to use a static reference of the json copy is not working out.  Thus can just perform a network query in an async task and just pass the return json to the data builder to get a populated set of recipes.

+++ app/src/main/java/com/example/nebo/bakingapp/data/Ingredient.java
@@ -1,7 +1,7 @@
 package com.example.nebo.bakingapp.data;
 
 public class Ingredient {
-    int quantity;
+    float quantity;
     String measure;
     String ingredient;
 }
